Moves Upgrades and Invulnerable trait to Mods.Common
