+++ OpenRA.Mods.Common/Traits/Upgrades/DisableUpgrade.cs
@@ -11,10 +11,9 @@
 using System;
 using System.Collections.Generic;
 using OpenRA.GameRules;
-using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Traits
 {
 	public class DisableUpgradeInfo : UpgradableTraitInfo, ITraitInfo
 	{
