+++ OpenRA.Mods.Common/Traits/Upgrades/UpgradeActorsNear.cs
@@ -11,10 +11,9 @@
 using System;
 using System.Collections.Generic;
 using OpenRA.GameRules;
-using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Traits
 {
 	[Desc("Applies an upgrade to actors within a specified range.")]
 	public class UpgradeActorsNearInfo : ITraitInfo
