+++ OpenRA.Mods.Common/Warheads/GrantUpgradeWarhead.cs
@@ -12,11 +12,11 @@ using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Effects;
 using OpenRA.GameRules;
+using OpenRA.Mods.Common.Effects;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Effects;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Warheads
 {
 	public class GrantUpgradeWarhead : Warhead
 	{
