+++ OpenRA.Mods.RA/Traits/InvulnerabilityUpgrade.cs
